+++ hazelcast-platform-labs-local/README.md
@@ -1,0 +1,3 @@
+This project is only used to package up a version of the labs into an uber-jar that includes hazelcast.  This is 
+useful when working locally but usually, when packaging a job for deployment, the Hazelcast classes should not be 
+included.
