--- src/main/resources/log4j.properties
@@ -1,6 +1,0 @@
-log4j.appender.stdout=org.apache.log4j.ConsoleAppender
-log4j.appender.stdout.Target=System.out
-log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
-log4j.appender.stdout.layout.ConversionPattern=%d{ABSOLUTE} [%c{1}] %m%n
-
-log4j.rootLogger=info, stdout
